requirements.txt
# ============================================
# VisionarySTEM - Dependencies
# ============================================

# --- Core AI ---
google-genai>=1.0.0
pydantic>=2.0

# --- Document Processing ---
PyMuPDF>=1.24.0

# --- Spatial RAG ---
chromadb>=0.5.0

# --- Text-to-Speech ---
edge-tts>=6.1.0

# --- API Server ---
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9

# --- Evaluation ---
jiwer>=3.0.0

# --- Config & Utils ---
python-dotenv>=1.0.0
aiofiles>=23.0.0
httpx>=0.27.0
